ci/jarvis-util/requirements.txt
pyyaml
pylint==2.15.0
coverage==5.5
# coverage
coverage-lcov==0.2.4
# coverage-lcov
pytest==6.2.5
pandas
tabulate